+++ frog-utils/src/ReactiveRichText/CustomQuillClipboard.js
