+++ frog-utils/src/index.js
@@ -48,10 +48,6 @@ export const ReactiveRichText = isBrowser
   ? require('./ReactiveRichText/main').default // eslint-disable-line global-require
   : () => <p>Node</p>; // React component to make Flow happy, will never be shown
 
-// export const LearningItemBlot = isBrowser
-// ? require('./LearningItemBlot').default // eslint-disable-line global-require
-// : () => <p>Node</p>; // React component to make Flow happy, will never be shown
-
 export {
   hideConditional,
   calculateHides,
